[fix] change domain level(add dictionary of garages and vehicles)

+++ Rental/models/dictionaries/vehicleTypes.cs
@@ -6,6 +6,6 @@ public class VehicleType
     public string Name { get; init; } = string.Empty;
     public string Category { get; init; } = string.Empty;
     public string Level { get; init; } = string.Empty;
-    public int Income { get; init; }
-    public int Price { get; init; }
+    public decimal Income { get; init; }
+    public decimal Price { get; init; }
 }
